+++ packages/database/CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.6] - 2025-05-04
+
+### Fixed
+
+- Fixed Prisma Studio connection URL
+
+## [1.0.5] - 2025-04-30
+
+### Changed
+
+- Only use SSL connection for connecting to AWS RDS
+
 ## [1.0.4] - 2026-04-29
 
 ### Removed